+++ src/components/Investment/InvestmentPatch.js
@@ -4,23 +4,14 @@ import visibilityOn from '../../assets/btn_visibility_on.svg';
 import visibilityOff from '../../assets/btn_visibility_off.svg';
 import Modal from '../Common/Modal';
 import { useState } from 'react';
-import { patchInvestment } from '../../api/InvestmentService';
 import InvestmentUpdate from './InvestmentUpdate';
 
 export default function InvestmentPatch({ onClose, mockInvestor, startup }) {
-  const {
-    id,
-    password: storedPassword,
-    ...investorDetails
-  } = mockInvestor || {};
+  const { password: storedPassword } = mockInvestor || {};
   const [password, setPassword] = useState('');
   const [isPasswordVisible, setIsPasswordVisible] = useState(false);
   const [showInvestmentUpdate, setShowInvestmentUpdate] = useState(false);
   const [error, setError] = useState('');
-
-  const [investAmount, setInvestAmount] = useState('');
-  const [comment, setComment] = useState('');
-  const [isPasswordValid, setIsPasswordValid] = useState(false);
 
   const togglePasswordVisibility = () => {
     setIsPasswordVisible(!isPasswordVisible);
@@ -32,9 +23,7 @@ export default function InvestmentPatch({ onClose, mockInvestor, startup }) {
 
   const handlePasswordSubmit = (e) => {
     e.preventDefault();
-    console.log('비밀번호 입력:', password);
-    console.log('저장된 비밀번호:', storedPassword);
-    console.log('타입:', typeof storedPassword);
+
     if (password !== storedPassword) {
       setError('비밀번호가 일치하지 않습니다.');
       console.log('비밀번호 불일치');
@@ -42,23 +31,6 @@ export default function InvestmentPatch({ onClose, mockInvestor, startup }) {
       setError('');
       setShowInvestmentUpdate(true);
       console.log('비밀번호 일치, 모달 열기');
-    }
-  };
-
-  const handleUpdateSubmit = async (e) => {
-    e.preventDefault();
-
-    try {
-      const updateRes = await patchInvestment(id, { investAmount });
-
-      if (updateRes.status === 200) {
-        onClose();
-      } else {
-        setError('수정 요청이 실패했습니다.');
-      }
-    } catch (err) {
-      console.error(err);
-      setError('투자 수정 중 오류가 발생했습니다.');
     }
   };
 
@@ -102,16 +74,17 @@ export default function InvestmentPatch({ onClose, mockInvestor, startup }) {
 
       {showInvestmentUpdate && (
         <InvestmentUpdate
-          onClick={handleUpdateSubmit}
           onClose={() => {
             setShowInvestmentUpdate(false);
             onClose();
           }}
           startup={startup}
+          mockInvestor={mockInvestor}
           initialValues={{
             name: mockInvestor.name,
             investAmount: mockInvestor.investAmount,
-            comment: mockInvestor.comment
+            comment: mockInvestor.comment,
+            password: mockInvestor.password
           }}
         />
       )}
